+++ src/components/color-picker/examples/color-picker-custom-palette.tsx
@@ -50,11 +50,7 @@ export class ColorPickerCustomPaletteExample {
                     onChange={this.onChange}
                     paletteColumnCount={this.paletteColumnCount}
                 />
-                <limel-example-controls
-                    style={{
-                        '--example-controls-column-layout': 'auto-fit',
-                    }}
-                >
+                <limel-example-controls>
                     <limel-checkbox
                         checked={this.paletteColumnCount === 3}
                         label="Set `paletteColumnCount` to 3"