+++ src/components/file/examples/file-composite.tsx
@@ -51,9 +51,7 @@ export class FileCompositeExample {
 
     private renderForm() {
         return (
-            <limel-example-controls
-                style={{ '--example-controls-column-layout': 'auto-fit' }}
-            >
+            <limel-example-controls>
                 <limel-form
                     schema={this.schema}
                     value={this.props}