+++ src/components/snackbar/examples/snackbar-persistent-non-dismissible.tsx
@@ -33,9 +33,7 @@ export class SnackbarPersistentNonDismissibleExample {
                 dismissible={false}
                 onHide={this.handleHide}
             />,
-            <limel-example-controls
-                style={{ '--example-controls-column-layout': 'auto-fit' }}
-            >
+            <limel-example-controls>
                 <limel-switch
                     label="Toggle online status"
                     onChange={this.handleChange}