+++ src/components/text-editor/examples/text-editor-ui.tsx
@@ -59,9 +59,7 @@ export class TextEditorUiExample {
                     ui={this.selectedUi.value}
                     placeholder={placeholderText}
                 />
-                <limel-example-controls
-                    style={{ '--example-controls-column-layout': 'auto-fit' }}
-                >
+                <limel-example-controls>
                     <limel-example-value value={this.value} />
                     <limel-select
                         label="ui"